smartinput_ng_daemon/pyproject.toml
[tool.poetry]
name = "smartinput_ng_daemon"
version = "0.1.0"
description = ""
authors = ["niwla23 <alwin@cloudserver.click>"]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
PyYAML = "^6.0"
requests = "^2.27.1"
PyAutoGUI = "^0.9.53"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
